requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==6.2.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
cryptography==45.0.5
distro==1.9.0
firebase_admin==7.0.0
firestore==0.0.8
Flask==3.1.2
flet==0.28.3
flet-desktop==0.28.3
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
msgpack==1.1.1
mypy_extensions==1.1.0
oauthlib==3.3.1
openai==2.8.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
proto-plus==1.26.1
protobuf==6.33.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
repath==0.9.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
yt-dlp==2025.11.12
